Use float instead of double for distance between vectors.

+++ DHClustering/KMVector.h
@@ -5,6 +5,6 @@
 
 + (id <KMVector>)meanOfVectors:(NSSet *)vectors;
 
-- (double)distanceFromVector:(id <KMVector>)otherVector;
+- (float)distanceFromVector:(id <KMVector>)otherVector;
 
 @end
